[refactor] 1. 配置项去除多余的字段 2. 等待时间单位统一修改为毫秒 [doc] 相关文档修改

+++ selenium-core/src/main/java/com/lml/selenium/dto/EleHandlerDto.java
@@ -48,8 +48,8 @@ public class EleHandlerDto extends BaseSeleniumDto {
     private Integer retry;
 
     /**
-     * 查询元素等待的时间(秒)
+     * 查询元素等待的时间(毫秒)
      */
-    private Integer waitTime;
+    private Long waitTime;
 
 }
